+++ Resources/Lang/en_US.json
@@ -2392,18 +2392,27 @@
   "RoleRateNoColor": "ON",
   "RoleOnNoColor": "ALWAYS",
   "RoleOffNoColor": "OFF",
-  "Rate0": "0%",
-  "Rate5": "5%",
-  "Rate10": "10%",
-  "Rate20": "20%",
-  "Rate30": "30%",
-  "Rate40": "40%",
-  "Rate50": "50%",
-  "Rate60": "60%",
-  "Rate70": "70%",
-  "Rate80": "80%",
-  "Rate90": "90%",
-  "Rate100": "100%",
+  "Rate0": "<#444444>0%</color>",
+  "Rate5": "<#808F8F>5%</color>",
+  "Rate10": "<#799595>10%</color>",
+  "Rate15": "<#719C9C>15%</color>",
+  "Rate20": "<#6AA2A2>20%</color>",
+  "Rate25": "<#62A9A9>25%</color>",
+  "Rate30": "<#5BB0B0>30%</color>",
+  "Rate35": "<#53B6B6>35%</color>",
+  "Rate40": "<#4CBDBD>40%</color>",
+  "Rate45": "<#44C4C4>45%</color>",
+  "Rate50": "<#3CCACA>50%</color>",
+  "Rate55": "<#35D1D1>55%</color>",
+  "Rate60": "<#2DD7D7>60%</color>",
+  "Rate65": "<#26DEDE>65%</color>",
+  "Rate70": "<#1EE5E5>70%</color>",
+  "Rate75": "<#17EBEB>75%</color>",
+  "Rate80": "<#0FF2F2>80%</color>",
+  "Rate85": "<#08F8F8>85%</color>",
+  "Rate90": "<#07F9F9>90%</color>",
+  "Rate95": "<#00FFFF>95%</color>",
+  "Rate100": "<#00FFA5>100%</color>",
   "Preset": "Preset",
   "Preset_1": "Preset 1",
   "Preset_2": "Preset 2",
