+++ Roles/Core/AssignManager/RoleAssign.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TOHE.Roles.Core.AssignManager;
+
+public static class RoleAssign
+{
+
+}
